+++ packages/amplify_api/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.2.7 (2021-11-08)
+
+### Chores
+
+- chore: Bump Amplify iOS to 1.15.5
+
+### Fixes
+
+- fix(api): Fix OIDC/Lambda in REST/GraphQL on Android
+
 ## 0.2.6 (2021-10-25)
 
 ## 0.2.5 (2021-10-14)